Add HTMX edit row example
I was the OP of this reddit post -
https://www.reddit.com/r/htmx/comments/16nui5k/concerned_hypersrcipt_could_be_the_barrier_to/
in which I argued that Hyperscript is offputting for people transitioning from
the Javascript world having one more thing to learn. In one of the
comments for the post, I was made aware of surreal. I thought it looked
excellent so I gave it a whirl by modifying the
[Edit Row example](https://htmx.org/examples/edit-row/) on the HTMX site
to work with surreal instead of hyperscript